+++ DeltaDNAAds/SmartAds/DDNASmartAdFactory.m
@@ -217,6 +217,11 @@ typedef NS_ENUM(NSInteger, DDNASmartAdAdapterType) {
                                                  configuration:configuration
                                                 waterfallIndex:i];
                 }
+                else if ([adProvider caseInsensitiveContains:@"TAPJOY"]) {
+                    adapter = [self instantiateAdapterForKlass:@"DDNASmartAdTapjoyAdapter"
+                                                 configuration:configuration
+                                                waterfallIndex:i];
+                }
                 else {
                     DDNALogWarn(@"Ad network %@ for %@ ads is not supported.",
                                 adProvider,