+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [Unreleased](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/1.5.0) (????-??-??)
+
+
 ## [1.4.5](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/1.4.5) (2017-06-21)
 ### Fixed
 - Not instantiating Facebook adapter.