+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/1.5.0) (????-??-??)
 ### Added
 - Send adSdkVersion when requesting configuration.
+- Support for Tapjoy.
 
 ## [1.4.5](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/1.4.5) (2017-06-21)
 ### Fixed